feat: add bull for iexec batch email send
