+++ .env.local.template
@@ -1,1 +1,6 @@
 NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID="1234";
+
+# Redis
+REDIS_HOST=localhost
+REDIS_PASSWORD=localhost
+REDIS_PORT=6379
