+++ package.json
@@ -71,6 +71,7 @@
     "@tanstack/react-query": "^5.20.2",
     "@tanstack/react-table": "^8.11.2",
     "@wagmi/core": "^2",
+    "bull": "^4.15.1",
     "chart.js": "^4.4.1",
     "chartjs-adapter-luxon": "^1.3.1",
     "class-variance-authority": "^0.7.0",
